Add correlation and distance metrics
After the disscusion on slack the used of KL-divergence is not really
suited for originality since if we shuffle a submission the distribution
will be the same and flag this as a false-positive.

The use of another (spearman) correlation metrics along side the pearson
correlation would be useful as this would detect any monotonic
correlation between two submissions.

Also added is using the mean canberra distance between the two distances
since it seems relatively robust against the cheating mentioned in the
PDF.