Change concordance to use test targets
Allowing concordance to be a measure of the percent difference in log
loss between the validation and test sets. This assumes that the test
set gathered from the competition data will have test targets. We are
only pulling a random 1/4 of the test set as a comparision against the
submission data. This is so that the "data leakage" mentioned in the pdf
is avoided.